refactor: consolidate tokenomics schema and utility functions into shared module, enhancing reusability and clarity across subgraphs
