+++ shared/tokenomics/schema-l2.graphql
