--- subgraphs/tokenomics-gnosis/schema.graphql
@@ -1,22 +1,0 @@
-type Token @entity(immutable: false) {
-  id: Bytes!
-  balance: BigInt!
-  holderCount: Int!
-}
-
-type TokenHolder @entity(immutable: false) {
-  id: Bytes!
-  token: Bytes!
-  balance: BigInt!
-}
-
-# Transfer event tracking for OLAS token
-type Transfer @entity(immutable: true) {
-  id: Bytes!
-  from: Bytes! # address
-  to: Bytes! # address
-  value: BigInt! # uint256
-  blockNumber: BigInt!
-  blockTimestamp: BigInt!
-  transactionHash: Bytes!
-}
